Xnurta_AMC/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
atomicwrites==1.4.1
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
execnet==2.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pefile==2023.2.7
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
pluggy==1.5.0
py ==1.11.0
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
PySocks==1.7.1
pytesseract==0.3.13
pytest~=8.3.5
pytest-html==3.1.1
pytest-metadata==3.1.1
pytest-parallel==0.0.10
pytest-xdist==3.6.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
selenium==4.29.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
tblib==3.0.0
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.0
typing_extensions==4.12.2
urllib3==2.3.0
urllib3-secure-extra==0.1.0
websocket-client==1.8.0
wsproto==1.2.0
zmail==0.2.8
pytest-rerunfailures==15.0
numpy~=2.2.4